"Checkpoint: Reevaluating certificate storage security."
The private certificate logic has been simplified but is not yet complete because I realized there is a security implication.
First thoughts and options have been laid out in '010_prepare_config_files.tf'.

Letting the options percolate for now but I'm favouring the use of an S3 bucket as the solution. Will spitball with schaluva to
get another opinion before deciding on the final course of action.