DOCS: Added root-level CHANGELOG.md
Honestly, this document should have existed several versions back (sorry!). Creating now to track the plethora of changes in the after-1.5.0 release (I assume we'll make it 2.0.0 due to breaking changes but TBD).

Document tries to capture a few things:
- Changes made to support Seqera Ecosystem directly.
- Changes made in CX Installer implementation (architecture / testing / etc)
- Tracking additions/changes/removal of keys from client interface points (terraform.tfvars & SSM secrets)

Initial focus is on getting this solid from the 1.5.0 -> new release. Once this is sorted out and people are given a chance to opine about my proposed approach, we will consider backporting changes from previous releases.