CHECKPOINT: Modified docker compose template (wave-lite conditional and labels)
The positive key YAML testing required me to change the container structure slightly. There was existing logic for when a key DOESN'T exist (expected value '') but this same flow doesn't work for
content when keys actually are present (since every keys content will be a complex structure with varying content).

To fix, I add a 'labels.seqera' key with a value that is the same as the docker compose service name. This provides a common way to consistently access structured information. We can infer
that the container is in docker compose by virtue of being able to retrieve its content, and the same is simple enough to standardize in tests without difficulty.

Like most of the testing framework, this feels kinda Rube Goldberg-y but I don't want to tear apart the testing framework right now, and the modification wasn't too onerous once I remembered that
positive YAML key validation works a tad differently than negative validation.