CHECKPOINT: Added configuration to control assignment of public IPv4 addresses to instances in public subnets.
While it is expected that most deployments will not need this feature (hence default to 'false'), I test often enough with an AWS Batch CE which leverages the same VPC as deployed by the Installer.
When I choose to use a public subnet, it always takes me a few minutes to realize that I've neglected to assign the IP, thus causing steps in the instance Launch Template (e.g. download Cloudwatch Agent
config) to fail because the HTTP traffic can't get back. Once this is realized the current instance must be terminated and a new one spun up. This wastes time, money, and is annoying.